first stab at refactoring
