+++ backend/app/services/evaluation/__init__.py
@@ -1,0 +1,32 @@
+"""Evaluation services."""
+
+from app.services.evaluation.dataset import upload_dataset
+from app.services.evaluation.evaluation import (
+    build_evaluation_config,
+    get_evaluation_with_scores,
+    start_evaluation,
+)
+from app.services.evaluation.validators import (
+    ALLOWED_EXTENSIONS,
+    ALLOWED_MIME_TYPES,
+    MAX_FILE_SIZE,
+    parse_csv_items,
+    sanitize_dataset_name,
+    validate_csv_file,
+)
+
+__all__ = [
+    # Dataset
+    "upload_dataset",
+    # Evaluation
+    "build_evaluation_config",
+    "get_evaluation_with_scores",
+    "start_evaluation",
+    # Validators
+    "ALLOWED_EXTENSIONS",
+    "ALLOWED_MIME_TYPES",
+    "MAX_FILE_SIZE",
+    "parse_csv_items",
+    "sanitize_dataset_name",
+    "validate_csv_file",
+]
